+++ inc/ti/fn/fnextendunique.h
@@ -41,7 +41,7 @@ static int do__f_extend_unique(ti_query_t * query, cleri_node_t * nd, ex_t * e)
             continue;
 
         ti_incref(v);
-        if (ti_varr_append(varr_dest, (void **) &v, e))
+        if (ti_val_varr_append(varr_dest, &v, e))
         {
             ti_decref(v);
             goto undo;