+++ inc/ti/fn/fnsplice.h
@@ -75,7 +75,7 @@ static int do__f_splice(ti_query_t * query, cleri_node_t * nd, ex_t * e)
         child = child->next->next;
 
         if (ti_do_statement(query, child->node, e) ||
-            ti_varr_append(varr, (void **) &query->rval, e))
+            ti_val_varr_append(varr, &query->rval, e))
             goto fail2;
 
         query->rval = NULL;